% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LQG_policy_recursion.R
\name{LQG_policy_recursion}
\alias{LQG_policy_recursion}
\title{Compute backward recursion to obtain the optimal policy}
\usage{
LQG_policy_recursion(RNderiv, transitions)
}
\arguments{
\item{RNderiv}{list with keys:
\code{A} 
\code{b}
\code{c}}

\item{transitions}{list with keys: 
\code{Kmats} array of size d x d x T 
\code{rvecs} array of size d x T
\code{Smats} array of size d x d x T}
}
\value{
list with keys: 
\code{As} 
\code{bs}
\code{cs}
}
\description{
Compute backward recursion to obtain the optimal policy, which has
the form at time t -log(psi_t) = x_t' A_t x_t + x_t'b_t + c_t.
}
