% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LQG_RNderiv.R
\name{LQG_RNderiv}
\alias{LQG_RNderiv}
\title{Obtain Radon Nikodym derivative}
\usage{
LQG_RNderiv(terminal, target)
}
\arguments{
\item{terminal}{list with keys:
\code{mean}
\code{inv.cov}}

\item{target}{list with keys: 
\code{mean}
\code{inv.cov}}
}
\value{
list with keys: 
\code{A} 
\code{b}
\code{c}
}
\description{
Obtain Radon Nikodym derivative between target and terminal distribution of current path measure. 
This has the form -log(RN) = x' A x + x'b + c.
}
