% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LQG_update_transitions.R
\name{LQG_update_transitions}
\alias{LQG_update_transitions}
\title{Update Markov transition kernels with new policy}
\usage{
LQG_update_transitions(policy_params, transitions)
}
\arguments{
\item{policy_params}{list with keys:
\code{As}
\code{bs}
\code{cs}}

\item{transitions}{list with keys: 
\code{Kmats} array of size d x d x T 
\code{rvecs} array of size d x T
\code{Smats} array of size d x d x T}
}
\value{
list with keys: 
\code{Kmats} array of size d x d x T 
\code{rvecs} array of size d x T
\code{Smats} array of size d x d x T
\code{inv.Smats} array of size d x d x T
}
\description{
# Function that takes the parameters produced by the policy recursion (LQG_policy_recursion)
and the parameters of the previous Gaussian transition kernels, and outputs the parameters
for the updated Gaussian transitions
}
