% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SchrodingerBridges.R
\docType{package}
\name{SchrodingerBridges}
\alias{SchrodingerBridges}
\alias{SchrodingerBridges-package}
\title{SchrodingerBridges package}
\description{
Numerical implementation of Schrodinger bridges
}
\details{
This package implements the iterated proportional fitting procedure to approximate Schrodinger bridges.
}
\author{
Jeremy Heng <jeremyhengjm@gmail.com>
}
\keyword{package}
